test: Handle 32 bit msec time overflows
The libinput_*_get_time() returns a 32 bit unsigned integer, but in the
tests we compared them to a 64 bit unsigned integer. This means that
when the 32 bit integer overflowed, we'd still compare to a
non-overflowed 64 bit integer, causing the tests to fail.

This commit fixes this by always casting the millisecond 64 bit unsigned
integer to a 32 unsigned integer, triggering the same overflow.

+++ test/touchpad.c
@@ -4017,7 +4017,7 @@ START_TEST(touchpad_time_usec)
 		utime = libinput_event_pointer_get_time_usec(ptrev);
 
 		ck_assert_int_eq(libinput_event_pointer_get_time(ptrev),
-				 utime / 1000);
+				 (uint32_t) (utime / 1000));
 		libinput_event_destroy(event);
 		event = libinput_get_event(li);
 	}
